[refactor] WatchRidingView 리펙토링
