+++ SnowMate/Shared/ModelContainer.swift
@@ -1,0 +1,8 @@
+//
+//  ModelContainer.swift
+//  SnowMateWatch Watch App
+//
+//  Created by 이혜빈 on 11/9/25.
+//
+
+import Foundation
