+++ SnowMate/SnowMateWatch Watch App/ContentView.swift
